# out/spring/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  description: An example customer portal OpenAPI specification
  title: Customer Portal API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /subscription:
    get:
      description: Get a list of all subscriptions
      operationId: getSubscription
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionsList'
          description: Successful response
      tags:
      - subscription
      x-accepts: application/json
      x-tags:
      - tag: subscription
    post:
      description: Create a new subscription
      operationId: addSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
        description: Update subscriber information
        required: true
      responses:
        "200":
          description: Successful response
      tags:
      - subscription
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: subscription
  /subscription/{subscriptionId}:
    delete:
      description: Update a subscription
      operationId: deleteSubscription
      parameters:
      - description: ID of subscription to delete
        explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          $ref: '#/components/schemas/SubscriptionID'
        style: simple
      responses:
        "200":
          description: Successful response
      tags:
      - subscription
      x-accepts: application/json
      x-tags:
      - tag: subscription
    get:
      description: Returns a subscription of a subscriber
      operationId: getSubscriptionById
      parameters:
      - description: ID of subscription to return
        explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          $ref: '#/components/schemas/SubscriptionID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
          description: Successful response
      tags:
      - subscription
      x-accepts: application/json
      x-tags:
      - tag: subscription
    put:
      description: Update a subscription
      operationId: updateSubscription
      parameters:
      - description: ID of subscription to update
        explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          $ref: '#/components/schemas/SubscriptionID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
        description: Update an existent pet in the store
        required: true
      responses:
        "200":
          description: Successful response
      tags:
      - subscription
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: subscription
  /subscription/{subscriptionId}/camera:
    get:
      description: Returns a list of all cameras that is within subscription
      operationId: getSubscribersCamera
      parameters:
      - description: ID of subscription to get the relevant cameras from
        explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          $ref: '#/components/schemas/SubscriptionID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CamerasList'
          description: Successful response
      tags:
      - subscription
      x-accepts: application/json
      x-tags:
      - tag: subscription
...
components:
  schemas:
...
    SubscriptionID:
      example: 311109
      format: int64
      type: integer
...
    Subscription:
      example:
        subscriberId: 311109
        id: 10
        camera:
          name: camera1
          location:
            latitude: 41.4
            longitude: 2.1
          id: 11
        status: active
      properties:
        id:
          example: 10
          format: int64
          type: integer
        subscriberId:
          example: 311109
          format: int64
          type: integer
        status:
          enum:
          - active
          - inactive
          - suspended
          - processing
          type: string
        camera:
          $ref: '#/components/schemas/Camera'
      type: object
    Camera:
      example:
        name: camera1
        location:
          latitude: 41.4
          longitude: 2.1
        id: 11
      properties:
        id:
          example: 11
          format: int64
          type: integer
        name:
          example: camera1
          type: string
        location:
          $ref: '#/components/schemas/Coordinates'
      required:
      - id
      type: object
    Coordinates:
      example:
        latitude: 41.4
        longitude: 2.1
      properties:
        latitude:
          example: 41.4
          format: double
          maximum: 90.0
          minimum: -90.0
          type: number
        longitude:
          example: 2.1
          format: double
          maximum: 180.0
          minimum: -180.0
          type: number
      required:
      - latitude
      - longitude
      type: object
...
    SubscriptionsList:
      items:
        $ref: '#/components/schemas/Subscription'
      type: array
    CamerasList:
      items:
        $ref: '#/components/schemas/Camera'
      type: array